backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib
cryptography
bcrypt